Move PusherData to ruma-common
